[cli] Move `dispatch` function to a separate file
This will make it easier to dispatch RPC aren't just for the main
command behavior (e.g. validating zone/instance names). The only code
differences here are that we pass the `Rpc::StubInterface` and `cerr`
stream to these new free functions.
